feat: modularize hero section and switch to external image URL
